Add Acceptance Tests
